add Doc for NllLoss

+++ onnx/defs/schema.cc
@@ -745,11 +745,11 @@ OpSchema& OpSchema::FillUsing(const std::function<void(OpSchema&)>& populator) {
 // only add input to the function body if it is used by a body's node.
 bool FilterFunctionBodyInputByName(const std::string& input_name, const FunctionProto& function_body)
 {
-  for (int i = 0; i< function_body.node_size(); i++)
+  for (size_t i = 0; i< function_body.node_size(); i++)
   {
-    auto node = function_body.node(i);
-    for (int j = 0; j < node.input_size(); j++)
-      if (node.input(j) == input_name)
+    auto node = function_body.node((int)i);
+    for (size_t j = 0; j < node.input_size(); j++)
+      if (node.input((int)j) == input_name)
         return true;
   }
   return false;
@@ -839,7 +839,7 @@ void OpSchema::Finalize() {
   if (this->HasFunction()) {
     BuildFunction(function_body_);
   } else {
-    for (int i = 0; i < queried_function_bodies_.size(); i++)
+    for (size_t i = 0; i < queried_function_bodies_.size(); i++)
       BuildFunction(queried_function_bodies_[i].second);
   }
 }
